S3Upload.Sign returns (string, string) instead of a struct.
This method will only ever return two strings out of it so there's no
reason to have a type specifically for the return type. The first string
is the base64-encoded policy document (which is passed in as an arg) and
the second string is the hex-encoded signature (which is calculated from
the policy document and other information contained in the S3Upload
struct).